+++ src/constants/constants.js
@@ -2,7 +2,12 @@ export const DefaultAppName = 'app';
 export const TemplateExt = '.mustache';
 export const ManifestoFileName = 'manifesto.json';
 export const DefaultPluralSuffix = 'Collection';
-export const Directive = { Copy: 'copy', Loop: 'loop', ViewModel: 'viewModel' };
+export const Directive = {
+  Copy: 'copy',
+  Loop: 'loop',
+  ViewModel: 'viewModel',
+  ViewModelTransform: 'viewModelTransform'
+};
 export const AllowedConfigProperties = [
   'name',
   'template',