+++ src/create/getViewModel.js
@@ -1,6 +1,15 @@
 import { getVariations } from '../lib/case.js';
 import { preprocessModels } from './viewModel/models.js';
 
+function addOffset(string, offset) {
+  const words = string.split('\n');
+  return words
+    .map((word) => {
+      return word.padStart(offset, ' ');
+    })
+    .join('\n');
+}
+
 export function getViewModel({ config, manifesto, dir }) {
   return {
     ...getVariations('appName', config.name),
@@ -8,6 +17,6 @@ export function getViewModel({ config, manifesto, dir }) {
     templateName: manifesto.name,
     templateVersion: manifesto.version,
     models: preprocessModels({ config, manifesto, dir }),
-    configString: JSON.stringify(config, null, 2)
+    configString: addOffset(JSON.stringify(config, null, 2), 2)
   };
 }