+++ src/create/viewModel/getDefaultValueForType.test.js
@@ -1,0 +1,9 @@
+import { getDefaultValueForType } from './getDefaultValueForType';
+
+describe('getDefaultValueForType', () => {
+  test('should handle missing type', async () => {
+    const actualDefaultValue = await getDefaultValueForType('test');
+
+    expect(actualDefaultValue).toEqual(null);
+  });
+});